api as a class itself
