+++ chia/_tests/core/server/test_server.py
@@ -16,7 +16,7 @@ from chia.protocols.full_node_protocol import RejectBlock, RequestBlock, Request
 from chia.protocols.protocol_message_types import ProtocolMessageTypes
 from chia.protocols.shared_protocol import Error, protocol_version
 from chia.protocols.wallet_protocol import RejectHeaderRequest
-from chia.server.api_protocol import ApiMethods
+from chia.server.api_protocol import ApiMetadata
 from chia.server.outbound_message import NodeType, make_msg
 from chia.server.server import ChiaServer
 from chia.server.start_full_node import create_full_node_service
@@ -25,7 +25,6 @@ from chia.server.ws_connection import WSChiaConnection, error_response_version
 from chia.simulator.block_tools import BlockTools
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.types.peer_info import PeerInfo
-from chia.util.api_decorators import api_request
 from chia.util.errors import ApiError, Err
 from chia.util.ints import int16, uint32
 
@@ -33,13 +32,13 @@ from chia.util.ints import int16, uint32
 @dataclass
 class TestAPI:
     log: logging.Logger = logging.getLogger(__name__)
-    api_methods: ClassVar[ApiMethods] = {}
+    api: ClassVar[ApiMetadata] = ApiMetadata()
 
     def ready(self) -> bool:
         return True
 
     # API call from FullNodeAPI
-    @api_request()
+    @api.request()
     async def request_transaction(self, request: RequestTransaction) -> None:
         raise ApiError(Err.NO_TRANSACTIONS_WHILE_SYNCING, f"Some error message: {request.transaction_id}", b"ab")
