+++ chia/data_layer/data_layer_api.py
@@ -4,14 +4,14 @@ import logging
 from typing import ClassVar
 
 from chia.data_layer.data_layer import DataLayer
-from chia.server.api_protocol import ApiMethods
+from chia.server.api_protocol import ApiMetadata
 from chia.server.server import ChiaServer
 
 
 class DataLayerAPI:
     log: logging.Logger
     data_layer: DataLayer
-    api_methods: ClassVar[ApiMethods] = {}
+    api: ClassVar[ApiMetadata] = ApiMetadata()
 
     def __init__(self, data_layer: DataLayer) -> None:
         self.log = logging.getLogger(__name__)
