+++ chia/introducer/introducer_api.py
@@ -7,19 +7,17 @@ from chia.introducer.introducer import Introducer
 from chia.protocols.introducer_protocol import RequestPeersIntroducer, RespondPeersIntroducer
 from chia.protocols.protocol_message_types import ProtocolMessageTypes
 from chia.rpc.rpc_server import StateChangedProtocol
-from chia.server.api_protocol import ApiMethods
+from chia.server.api_protocol import ApiMetadata
 from chia.server.outbound_message import Message, make_msg
 from chia.server.ws_connection import WSChiaConnection
 from chia.types.peer_info import TimestampedPeerInfo
-from chia.util.api_decorators import api_request, collect_api_methods
 from chia.util.ints import uint64
 
 
-@collect_api_methods
 class IntroducerAPI:
     log: logging.Logger
     introducer: Introducer
-    api_methods: ClassVar[ApiMethods] = {}
+    api: ClassVar[ApiMetadata] = ApiMetadata()
 
     def __init__(self, introducer) -> None:
         self.log = logging.getLogger(__name__)
@@ -31,7 +29,7 @@ class IntroducerAPI:
     def _set_state_changed_callback(self, callback: StateChangedProtocol) -> None:
         pass
 
-    @api_request(peer_required=True)
+    @api.request(peer_required=True)
     async def request_peers_introducer(
         self,
         request: RequestPeersIntroducer,
