+++ chia/wallet/nft_wallet/nft_wallet.py
@@ -11,6 +11,7 @@ from chia_rs import AugSchemeMPL, G1Element, G2Element
 from clvm.casts import int_from_bytes, int_to_bytes
 from typing_extensions import Unpack
 
+import chia.server.api_protocol
 import chia.wallet.singleton
 from chia.protocols.wallet_protocol import CoinState
 from chia.server.ws_connection import WSChiaConnection
@@ -94,7 +95,7 @@ class NFTWallet:
         self.standard_wallet = wallet
         if name is None:
             name = "NFT Wallet"
-        self.log = logging.getLogger(name if name else __name__)
+        chia.server.api_protocol.log = logging.getLogger(name if name else __name__)
         self.wallet_state_manager = wallet_state_manager
         self.nft_wallet_info = NFTWalletInfo(did_id)
         info_as_string = json.dumps(self.nft_wallet_info.to_json_dict())
@@ -103,10 +104,12 @@ class NFTWallet:
         )
         self.wallet_id = self.wallet_info.id
         self.nft_store = wallet_state_manager.nft_store
-        self.log.debug("NFT wallet id: %r and standard wallet id: %r", self.wallet_id, self.standard_wallet.wallet_id)
+        chia.server.api_protocol.log.debug(
+            "NFT wallet id: %r and standard wallet id: %r", self.wallet_id, self.standard_wallet.wallet_id
+        )
 
         await self.wallet_state_manager.add_new_wallet(self)
-        self.log.debug("Generated a new NFT wallet: %s", self.__dict__)
+        chia.server.api_protocol.log.debug("Generated a new NFT wallet: %s", self.__dict__)
         return self
 
     @classmethod
@@ -118,7 +121,7 @@ class NFTWallet:
         name: Optional[str] = None,
     ) -> _T_NFTWallet:
         self = cls()
-        self.log = logging.getLogger(name if name else __name__)
+        chia.server.api_protocol.log = logging.getLogger(name if name else __name__)
         self.wallet_state_manager = wallet_state_manager
         self.wallet_info = wallet_info
         self.wallet_id = wallet_info.id
