+++ chia/wallet/wallet_node_api.py
@@ -4,20 +4,18 @@ import logging
 from typing import ClassVar
 
 from chia.protocols import full_node_protocol, introducer_protocol, wallet_protocol
-from chia.server.api_protocol import ApiMethods
+from chia.server.api_protocol import ApiMetadata
 from chia.server.outbound_message import NodeType
 from chia.server.ws_connection import WSChiaConnection
 from chia.types.mempool_inclusion_status import MempoolInclusionStatus
-from chia.util.api_decorators import api_request, collect_api_methods
 from chia.util.errors import Err
 from chia.wallet.wallet_node import WalletNode
 
 
-@collect_api_methods
 class WalletNodeAPI:
     log: logging.Logger
     wallet_node: WalletNode
-    api_methods: ClassVar[ApiMethods] = {}
+    api: ClassVar[ApiMetadata] = ApiMetadata()
 
     def __init__(self, wallet_node) -> None:
         self.log = logging.getLogger(__name__)
@@ -26,7 +24,7 @@ class WalletNodeAPI:
     def ready(self) -> bool:
         return self.wallet_node.logged_in
 
-    @api_request(peer_required=True)
+    @api.request(peer_required=True)
     async def respond_removals(self, response: wallet_protocol.RespondRemovals, peer: WSChiaConnection):
         pass
 
@@ -36,14 +34,14 @@ class WalletNodeAPI:
         """
         pass
 
-    @api_request()
+    @api.request()
     async def reject_additions_request(self, response: wallet_protocol.RejectAdditionsRequest):
         """
         The full node has rejected our request for additions.
         """
         pass
 
-    @api_request(peer_required=True, execute_task=True)
+    @api.request(peer_required=True, execute_task=True)
     async def new_peak_wallet(self, peak: wallet_protocol.NewPeakWallet, peer: WSChiaConnection):
         """
         The full node sent as a new peak
@@ -73,26 +71,26 @@ class WalletNodeAPI:
 
         await self.wallet_node.new_peak_queue.new_peak_wallet(peak, peer)
 
-    @api_request()
+    @api.request()
     async def reject_header_request(self, response: wallet_protocol.RejectHeaderRequest):
         """
         The full node has rejected our request for a header.
         """
         pass
 
-    @api_request()
+    @api.request()
     async def respond_block_header(self, response: wallet_protocol.RespondBlockHeader):
         pass
 
-    @api_request(peer_required=True)
+    @api.request(peer_required=True)
     async def respond_additions(self, response: wallet_protocol.RespondAdditions, peer: WSChiaConnection):
         pass
 
-    @api_request()
+    @api.request()
     async def respond_proof_of_weight(self, response: full_node_protocol.RespondProofOfWeight):
         pass
 
-    @api_request(peer_required=True)
+    @api.request(peer_required=True)
     async def transaction_ack(self, ack: wallet_protocol.TransactionAck, peer: WSChiaConnection):
         """
         This is an ack for our previous SendTransaction call. This removes the transaction from
@@ -132,7 +130,7 @@ class WalletNodeAPI:
             else:
                 await wallet_state_manager.remove_from_queue(ack.txid, name, status, None)
 
-    @api_request(peer_required=True)
+    @api.request(peer_required=True)
     async def respond_peers_introducer(
         self, request: introducer_protocol.RespondPeersIntroducer, peer: WSChiaConnection
     ):
@@ -142,7 +140,7 @@ class WalletNodeAPI:
         if peer is not None and peer.connection_type is NodeType.INTRODUCER:
             await peer.close()
 
-    @api_request(peer_required=True)
+    @api.request(peer_required=True)
     async def respond_peers(self, request: full_node_protocol.RespondPeers, peer: WSChiaConnection):
         if self.wallet_node.wallet_peers is None:
             return None
@@ -152,54 +150,54 @@ class WalletNodeAPI:
 
         return None
 
-    @api_request()
+    @api.request()
     async def respond_puzzle_solution(self, request: wallet_protocol.RespondPuzzleSolution):
         self.log.error("Unexpected message `respond_puzzle_solution`. Peer might be slow to respond")
         return None
 
-    @api_request()
+    @api.request()
     async def reject_puzzle_solution(self, request: wallet_protocol.RejectPuzzleSolution):
         self.log.warning(f"Reject puzzle solution: {request}")
 
-    @api_request()
+    @api.request()
     async def respond_header_blocks(self, request: wallet_protocol.RespondHeaderBlocks):
         pass
 
-    @api_request()
+    @api.request()
     async def respond_block_headers(self, request: wallet_protocol.RespondBlockHeaders):
         pass
 
-    @api_request()
+    @api.request()
     async def reject_header_blocks(self, request: wallet_protocol.RejectHeaderBlocks):
         self.log.warning(f"Reject header blocks: {request}")
 
-    @api_request()
+    @api.request()
     async def reject_block_headers(self, request: wallet_protocol.RejectBlockHeaders):
         pass
 
-    @api_request(peer_required=True, execute_task=True)
+    @api.request(peer_required=True, execute_task=True)
     async def coin_state_update(self, request: wallet_protocol.CoinStateUpdate, peer: WSChiaConnection):
         await self.wallet_node.new_peak_queue.full_node_state_updated(request, peer)
 
     # TODO: Review this hinting issue around this rust type not being a Streamable
     #       subclass, as you might expect it wouldn't be.  Maybe we can get the
     #       protocol working right back at the api_request definition.
-    @api_request()  # type: ignore[type-var]
+    @api.request()  # type: ignore[type-var]
     async def respond_to_ph_update(self, request: wallet_protocol.RespondToPhUpdates):
         pass
 
-    @api_request()
+    @api.request()
     async def respond_to_coin_update(self, request: wallet_protocol.RespondToCoinUpdates):
         pass
 
-    @api_request()
+    @api.request()
     async def respond_children(self, request: wallet_protocol.RespondChildren):
         pass
 
-    @api_request()
+    @api.request()
     async def respond_ses_hashes(self, request: wallet_protocol.RespondSESInfo):
         pass
 
-    @api_request()
+    @api.request()
     async def respond_blocks(self, request: full_node_protocol.RespondBlocks) -> None:
         pass
